Clean up js; Reorganize controllers
