+++ src/js/controllers/labs.js
