+++ src/js/controllers/markets/view.js
