+++ src/js/controllers/messages/index.js
@@ -1,0 +1,9 @@
+module.exports = [
+  '$scope',
+  '$location',
+  function ($scope, $location) {
+    $scope.open = function (message) {
+      $location.path('/messages/' + message.id);
+    };
+  }
+];
