+++ src/js/controllers/messages/view.js
@@ -1,0 +1,5 @@
+module.exports = [
+  '$scope',
+  function ($scope) {
+  }
+];
