+++ src/js/controllers/settings.js
